+++ src/renderer.js
@@ -10,25 +10,25 @@ import * as utils from './utils';
 import i18next from 'i18next';
 
 function createVdom(node) {
-  let virtualizeTime = new Instrument();
+  const virtualizeTime = new Instrument();
   virtualizeTime.start();
 
-  let vNode = virtualize(node);
+  const vNode = virtualize(node);
 
-  i18next.services.logger.log('virtualization took: ' + virtualizeTime.end() + 'ms');
+  i18next.services.logger.log(`virtualization took: ${virtualizeTime.end()}ms`);
   return vNode;
 }
 
-export default function(root, observer) {
-  let ret = {};
-  ret.render = function render() {
-    let newNode = createVdom(root);
-    let localized = localize(ultraDeepClone(newNode));
+export default function (root, observer) {
+  const ret = {};
+  ret.render = function render(retranslate) {
+    const newNode = createVdom(root);
+    const localized = localize(ultraDeepClone(newNode), retranslate);
 
-    let patches = diff(newNode, localized);
+    const patches = diff(newNode, localized);
     if (patches['0']) observer.reset(); // reset observer if having patches
     root = patch(root, patches);
-  }
+  };
 
   ret.debouncedRender = utils.debounce(ret.render, 200);